requirements.txt
anthropic
chromadb
pandas
pdfplumber
sentence-transformers
python-dotenv
requests
numpy
streamlit